Move classes out of the same file - Added OwnedNativeMemory and OwnedPinnedArray class files
